Extract shared ActivityDetailModal component
The activity detail popup (icon/label with subtypes, logged time + byline,
duration, notes, optional Open-Journal link) was duplicated across the
dashboard, caretaker, journal index, journal day, and TodayEventsList. Pull it
into one full-modal component; sites pass an optional journalHref or a footer
snippet for their per-context action. Unifies drifted label keys, footer
icon/label, and aria strings onto the dashboard's treatment.
