+++ src/routes/(app)/(companion)/[companionId]/journal/[date]/+page.svelte
@@ -25,15 +25,13 @@
 		Plus,
 		Pencil,
 		NotebookPen,
-		X,
 		Activity
 	} from '@lucide/svelte';
 	import EmptyState from '$lib/components/EmptyState.svelte';
 	import SubtypePills from '$lib/components/log/SubtypePills.svelte';
+	import ActivityDetailModal from '$lib/components/log/ActivityDetailModal.svelte';
 	import { Button } from '$lib/components/ui/button/index.js';
-	import { Separator } from '$lib/components/ui/separator/index.js';
 	import { Badge } from '$lib/components/ui/badge/index.js';
-	import LocalTime from '$lib/components/LocalTime.svelte';
 	import ByLine from '$lib/components/ByLine.svelte';
 	import { SvelteDate } from 'svelte/reactivity';
 	import { localDatetimes } from '$lib/actions/localDatetimes';
@@ -339,10 +337,6 @@
 	});
 
 	function handleKeydown(e: KeyboardEvent) {
-		if (e.key === 'Escape' && detailEvent) {
-			closeActivityDetail();
-			return;
-		}
 		if ((e.metaKey || e.ctrlKey) && e.key === 'p') {
 			e.preventDefault();
 			viewMode = viewMode === 'write' ? 'preview' : 'write';
@@ -398,43 +392,15 @@
 	let deleteActivityId = $state('');
 	let deleteActivityForm = $state<HTMLFormElement | null>(null);
 
-	// Activity detail modal
+	// Activity detail modal (shared component owns focus/escape/backdrop).
 	let detailEvent = $state<(typeof data.dailyEvents)[0] | null>(null);
-	let detailDialogEl = $state<HTMLElement | null>(null);
 
-	async function openActivityDetail(event: (typeof data.dailyEvents)[0]) {
+	function openActivityDetail(event: (typeof data.dailyEvents)[0]) {
 		detailEvent = event;
-		await tick();
-		detailDialogEl?.focus();
 	}
 
 	function closeActivityDetail() {
 		detailEvent = null;
-	}
-
-	function trapDetailFocus(e: KeyboardEvent) {
-		if (!detailDialogEl) return;
-		const focusable = Array.from(
-			detailDialogEl.querySelectorAll<HTMLElement>(
-				'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])'
-			)
-		).filter((el) => !el.hasAttribute('disabled'));
-		if (!focusable.length) return;
-		const first = focusable[0];
-		const last = focusable[focusable.length - 1];
-		if (e.key === 'Tab') {
-			if (e.shiftKey) {
-				if (document.activeElement === first) {
-					e.preventDefault();
-					last.focus();
-				}
-			} else {
-				if (document.activeElement === last) {
-					e.preventDefault();
-					first.focus();
-				}
-			}
-		}
 	}
 </script>
 
@@ -446,96 +412,22 @@
 
 <!-- Activity detail modal -->
 {#if detailEvent}
-	<div class="fixed inset-0 z-50 flex items-end sm:items-center justify-center p-4 sm:p-6">
-		<button
-			tabindex="-1"
-			class="absolute inset-0 bg-black/50 backdrop-blur-sm"
-			aria-label={t(locale, 'aria.closeDialog')}
-			onclick={closeActivityDetail}
-		></button>
-		<div
-			bind:this={detailDialogEl}
-			role="dialog"
-			aria-modal="true"
-			tabindex="-1"
-			onkeydown={trapDetailFocus}
-			class="relative z-10 w-full max-w-md rounded-xl border bg-card text-card-foreground shadow-xl focus:outline-none
-				animate-in fade-in-0 zoom-in-95 slide-in-from-bottom-4 sm:slide-in-from-bottom-0 duration-200"
-		>
-			<div class="flex items-center justify-between px-5 pt-5 pb-3">
-				<h2 class="font-semibold text-base text-foreground">
-					{activityDisplayIcon(detailEvent.type, detailEvent.subtypes)}
-					{activityDisplayLabel(locale, detailEvent.type, detailEvent.subtypes)}
-				</h2>
-				<button
-					onclick={closeActivityDetail}
-					aria-label={t(locale, 'aria.close')}
-					class="rounded-md p-1 text-muted-foreground hover:text-foreground hover:bg-accent transition-colors"
-				>
-					<X class="h-4 w-4" />
-				</button>
-			</div>
-
-			<Separator />
-
-			<div class="px-5 py-4 space-y-3 text-sm">
-				<div class="flex items-center gap-3">
-					<span class="w-20 shrink-0 text-xs font-medium text-muted-foreground"
-						>{t(locale, 'page.journal.day.detailType')}</span
-					>
-					<Badge variant="gold"
-						>{activityDisplayLabel(locale, detailEvent.type, detailEvent.subtypes)}</Badge
-					>
-				</div>
-				<div class="flex items-center gap-3">
-					<span class="w-20 shrink-0 text-xs font-medium text-muted-foreground"
-						>{t(locale, 'page.journal.day.detailLogged')}</span
-					>
-					<span class="text-foreground"
-						><LocalTime date={detailEvent.loggedAt} format="datetime" /><ByLine
-							user={detailEvent.logger}
-							variant="inline"
-						/></span
-					>
-				</div>
-				{#if detailEvent.durationMinutes}
-					<div class="flex items-center gap-3">
-						<span class="w-20 shrink-0 text-xs font-medium text-muted-foreground"
-							>{t(locale, 'page.journal.day.detailDuration')}</span
-						>
-						<span class="text-foreground">{detailEvent.durationMinutes} min</span>
-					</div>
-				{/if}
-				{#if detailEvent.notes}
-					<div class="pt-1">
-						<p class="text-xs font-medium text-muted-foreground mb-1">
-							{t(locale, 'page.journal.day.detailNotes')}
-						</p>
-						<div class="prose prose-sm dark:prose-invert max-w-none">
-							{@html renderMarkdown(detailEvent.notes)}
-						</div>
-					</div>
-				{/if}
-			</div>
-
-			<Separator />
-
-			<div class="flex gap-2 px-5 py-4">
-				<Button
-					variant="soft"
-					size="sm"
-					onclick={() => {
-						const ev = detailEvent;
-						closeActivityDetail();
-						if (ev) startEditActivity(ev);
-					}}
-				>
-					<Pencil class="h-3.5 w-3.5 mr-1.5" />
-					{t(locale, 'common.edit')}
-				</Button>
-			</div>
-		</div>
-	</div>
+	<ActivityDetailModal event={detailEvent} onclose={closeActivityDetail}>
+		{#snippet footer()}
+			<Button
+				variant="soft"
+				size="sm"
+				onclick={() => {
+					const ev = detailEvent;
+					closeActivityDetail();
+					if (ev) startEditActivity(ev);
+				}}
+			>
+				<Pencil class="h-3.5 w-3.5 mr-1.5" />
+				{t(locale, 'common.edit')}
+			</Button>
+		{/snippet}
+	</ActivityDetailModal>
 {/if}
 
 <div class="max-w-3xl mx-auto pb-24 md:pb-0">
